Initializiation
Initialization for the html and css, this was done first because:

I already have most of the backend because I'm redoing the challange because I was not happy with the layout that I had.
For that reason I started prototyping what I actually wanted from the front end, which is just easier to do when you don't have other code messing it up.